+++ src/main/java/org/hihan/girinoscope/ui/HtmlPane.java
@@ -11,8 +11,11 @@ import java.net.URISyntaxException;
 import java.net.URL;
 import java.util.Enumeration;
 import java.util.Objects;
+import java.util.Properties;
 import java.util.logging.Level;
 import java.util.logging.Logger;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 import javax.swing.JEditorPane;
 import javax.swing.event.HyperlinkEvent;
 import javax.swing.event.HyperlinkListener;
@@ -36,7 +39,7 @@ public class HtmlPane extends JEditorPane {
     }
 
     private static String loadContent(URL url) throws IOException {
-        try (InputStream input = url.openStream()) {
+        try ( InputStream input = url.openStream()) {
             BufferedReader reader = new BufferedReader(new InputStreamReader(input));
             StringBuilder file = new StringBuilder();
             String line;
@@ -47,8 +50,23 @@ public class HtmlPane extends JEditorPane {
         }
     }
 
+    private static String expand(String text, Properties properties) {
+        Pattern variablePattern = Pattern.compile("\\$\\{([A-Za-z0-9.\\-_]+)\\}");
+        Matcher matcher = variablePattern.matcher(text);
+        StringBuffer expandedText = new StringBuffer();
+        while (matcher.find()) {
+            matcher = matcher.appendReplacement(expandedText, properties.getProperty(matcher.group(1)));
+        }
+        matcher.appendTail(expandedText);
+        return expandedText.toString();
+    }
+
     public HtmlPane(URL url) throws IOException {
         this(loadContent(url));
+    }
+
+    public HtmlPane(URL url, Properties properties) throws IOException {
+        this(expand(loadContent(url), properties));
     }
 
     public HtmlPane(String text) {
@@ -69,19 +87,24 @@ public class HtmlPane extends JEditorPane {
         StyleSheet styleSheet = ((HTMLDocument) super.getDocument()).getStyleSheet();
         styleSheet.addRule("body {bgcolor: white; font-family: Sans-Serif;}");
 
-        super.addHyperlinkListener(new HyperlinkListener() {
-            @Override
-            public void hyperlinkUpdate(final HyperlinkEvent event) {
-                if (event.getEventType() == HyperlinkEvent.EventType.ACTIVATED) {
-                    final String href = getHref(event);
-                    try {
-                        Desktop.getDesktop().browse(new URI(href));
-                    } catch (URISyntaxException | IOException e) {
-                        LOGGER.log(Level.WARNING, "Can’t open link " + href, e);
+        if (Desktop.isDesktopSupported()) {
+            final Desktop desktop = Desktop.getDesktop();
+            if (desktop.isSupported(Desktop.Action.BROWSE)) {
+                super.addHyperlinkListener(new HyperlinkListener() {
+                    @Override
+                    public void hyperlinkUpdate(final HyperlinkEvent event) {
+                        if (event.getEventType() == HyperlinkEvent.EventType.ACTIVATED) {
+                            final String href = getHref(event);
+                            try {
+                                desktop.browse(new URI(href));
+                            } catch (URISyntaxException | IOException e) {
+                                LOGGER.log(Level.WARNING, "Can’t open link " + href, e);
+                            }
+                        }
                     }
-                }
+                });
             }
-        });
+        }
     }
 
     private static String getHref(HyperlinkEvent event) {